requirements-minimal.txt
# requirements-minimal.txt - Versión ligera sin marker-pdf
# Para usuarios que solo quieren usar python-docx para procesamiento de documentos

# --- Core ML y NLP ---
transformers>=4.36.0
sentence-transformers>=2.2.2
numpy>=1.24.0

# --- Procesamiento de Documentos (Solo python-docx) ---
python-docx>=1.1.0             # Procesamiento básico de DOCX
langchain-text-splitters>=0.0.1

# --- Base de Datos Vectorial ---
psycopg2-binary>=2.9.7         # Cliente PostgreSQL
pgvector>=0.2.3                # Extensión vectorial para PostgreSQL

# --- Interfaz Web ---
streamlit>=1.28.0              # Interfaz web interactiva

# --- APIs y Cliente LLM ---
openai>=1.3.0                  # Cliente OpenAI compatible con vLLM
requests>=2.31.0               # Llamadas HTTP a endpoints vLLM

# --- Búsqueda Textual ---
lunr>=0.6.2                    # Índice de búsqueda textual

# --- Configuración ---
python-dotenv>=1.0.0           # Manejo de variables de entorno

# --- Backend para modelos (mínimo) ---
torch>=2.0.0                   # Backend para sentence-transformers 